Set datamodule transforms property from preprocessor
